+++ tests/api/PDFDocument.spec.ts
@@ -12,23 +12,23 @@ import type {
 import {
   Duplex,
   NonFullScreenPageMode,
-  PDFArray,
-  PDFDict,
-  PDFHexString,
-  PDFName,
-  PDFRef,
   PrintScaling,
   ReadingDirection,
   ViewerPreferences,
-} from 'src/core/index.js';
+} from 'src/core/interactive/ViewerPreferences.js';
+import { PDFArray } from 'src/core/objects/PDFArray.js';
+import { PDFDict } from 'src/core/objects/PDFDict.js';
+import { PDFHexString } from 'src/core/objects/PDFHexString.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
 import {
   EncryptedPDFError,
   InvalidFontSubsetOptionError,
-  ParseSpeeds,
-  PDFDocument,
-  PDFPage,
-  PDFFont,
-} from 'src/api/index.js';
+} from 'src/api/errors.js';
+import { ParseSpeeds } from 'src/api/PDFDocumentOptions.js';
+import { PDFDocument } from 'src/api/PDFDocument.js';
+import { PDFPage } from 'src/api/PDFPage.js';
+import { PDFFont } from 'src/api/PDFFont.js';
 import {
   InvalidIndirectObjectError,
   FontkitAssertionError as PDFLibFontkitAssertionError,
@@ -36,7 +36,7 @@ import {
 } from 'src/core/errors.js';
 import { PDFSecurity, SecurityOptions } from 'src/core/security/PDFSecurity.js';
 import { readBinaryFileSync } from '../test-utils.js';
-import { InvalidPngError } from 'src/utils/index.js';
+import { InvalidPngError } from 'src/utils/errors.js';
 
 const examplePngImage = readBinaryFileSync('assets/images/etwe.png');
