+++ tests/api/PDFImage.spec.ts
@@ -1,5 +1,6 @@
-import { PDFDocument, PDFImage } from 'src/api/index.js';
-import { PngEmbedder } from 'src/core/index.js';
+import { PDFDocument } from 'src/api/PDFDocument.js';
+import { PDFImage } from 'src/api/PDFImage.js';
+import { PngEmbedder } from 'src/core/embedders/PngEmbedder.js';
 import { readBinaryFileSync } from '../test-utils.js';
 
 const examplePngImage = readBinaryFileSync('assets/images/etwe.png');
