+++ tests/api/form/PDFCheckBox.spec.ts
@@ -1,5 +1,5 @@
-import { AnnotationFlags } from 'src/core/index.js';
-import { PDFDocument } from 'src/api/index.js';
+import { AnnotationFlags } from 'src/core/annotation/flags.js';
+import { PDFDocument } from 'src/api/PDFDocument.js';
 import { readBinaryFileSync } from '../../test-utils.js';
 
 const fancyFieldsPdfBytes = readBinaryFileSync('assets/pdfs/fancy_fields.pdf');
