+++ tests/api/form/PDFForm.spec.ts
@@ -1,20 +1,16 @@
-import {
-  PDFAcroForm,
-  PDFDict,
-  PDFRef,
-  PDFName,
-  PDFWidgetAnnotation,
-} from 'src/core/index.js';
-import {
-  PDFDocument,
-  PDFTextField,
-  PDFCheckBox,
-  PDFButton,
-  PDFRadioGroup,
-  PDFOptionList,
-  PDFDropdown,
-  PDFForm,
-} from 'src/api/index.js';
+import { PDFAcroForm } from 'src/core/acroform/PDFAcroForm.js';
+import { PDFWidgetAnnotation } from 'src/core/annotation/PDFWidgetAnnotation.js';
+import { PDFDict } from 'src/core/objects/PDFDict.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { PDFButton } from 'src/api/form/PDFButton.js';
+import { PDFCheckBox } from 'src/api/form/PDFCheckBox.js';
+import { PDFDropdown } from 'src/api/form/PDFDropdown.js';
+import { PDFForm } from 'src/api/form/PDFForm.js';
+import { PDFOptionList } from 'src/api/form/PDFOptionList.js';
+import { PDFRadioGroup } from 'src/api/form/PDFRadioGroup.js';
+import { PDFTextField } from 'src/api/form/PDFTextField.js';
+import { PDFDocument } from 'src/api/PDFDocument.js';
 import { readBinaryFileSync } from '../../test-utils.js';
 
 const getWidgets = (pdfDoc: PDFDocument) =>
