+++ tests/api/form/PDFTextField.spec.ts
@@ -1,5 +1,6 @@
-import { AnnotationFlags } from 'src/core/index.js';
-import { PDFDocument, TextAlignment } from 'src/api/index.js';
+import { AnnotationFlags } from 'src/core/annotation/flags.js';
+import { PDFDocument } from 'src/api/PDFDocument.js';
+import { TextAlignment } from 'src/api/text/alignment.js';
 import { readBinaryFileSync } from '../../test-utils.js';
 
 const fancyFieldsPdfBytes = readBinaryFileSync('assets/pdfs/fancy_fields.pdf');
