+++ tests/core/PDFContext.spec.ts
@@ -1,19 +1,17 @@
 import pako from 'pako';
 
-import {
-  PDFArray,
-  PDFBool,
-  PDFContentStream,
-  PDFContext,
-  PDFDict,
-  PDFHexString,
-  PDFName,
-  PDFNull,
-  PDFNumber,
-  PDFRef,
-  PDFString,
-} from 'src/core/index.js';
-import { mergeIntoTypedArray } from 'src/utils/index.js';
+import { PDFContext } from 'src/core/PDFContext.js';
+import { PDFArray } from 'src/core/objects/PDFArray.js';
+import { PDFBool } from 'src/core/objects/PDFBool.js';
+import { PDFDict } from 'src/core/objects/PDFDict.js';
+import { PDFHexString } from 'src/core/objects/PDFHexString.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFNull } from 'src/core/objects/PDFNull.js';
+import { PDFNumber } from 'src/core/objects/PDFNumber.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { PDFString } from 'src/core/objects/PDFString.js';
+import { PDFContentStream } from 'src/core/structures/PDFContentStream.js';
+import { mergeIntoTypedArray } from 'src/utils/arrays.js';
 
 describe(`PDFContext`, () => {
   it(`retains assigned objects`, () => {
