+++ tests/core/document/PDFHeader.spec.ts
@@ -1,5 +1,6 @@
-import { PDFHeader } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PDFHeader } from 'src/core/document/PDFHeader.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 
 describe(`PDFHeader`, () => {
   it(`can be constructed from PDFHeader.forVersion(...)`, () => {
