+++ tests/core/document/PDFTrailerDict.spec.ts
@@ -1,5 +1,7 @@
-import { PDFContext, PDFTrailerDict } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PDFContext } from 'src/core/PDFContext.js';
+import { PDFTrailerDict } from 'src/core/document/PDFTrailerDict.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 
 describe(`PDFTrailerDict`, () => {
   const dict = PDFContext.create().obj({ Foo: 'Bar' });
