+++ tests/core/embedders/CMap.spec.ts
@@ -3,7 +3,7 @@ import type { TTFFont, Glyph } from '@denkiyagi/fontkit';
 import fs from 'fs';
 
 import { createCmap } from 'src/core/embedders/CMap.js';
-import { byAscendingId, sortedUniq } from 'src/utils/index.js';
+import { byAscendingId, sortedUniq } from 'src/utils/arrays.js';
 
 const ubuntuFont = fs.readFileSync('./assets/fonts/ubuntu/Ubuntu-R.ttf');
 const sourceHansJpFont = fs.readFileSync(
