+++ tests/core/objects/PDFBool.spec.ts
@@ -1,5 +1,8 @@
-import { PDFBool, PDFRef, PrivateConstructorError } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PrivateConstructorError } from 'src/core/errors.js';
+import { PDFBool } from 'src/core/objects/PDFBool.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 import { security } from './shared.js';
 
 describe(`PDFBool`, () => {
