+++ tests/core/objects/PDFName.spec.ts
@@ -1,5 +1,8 @@
-import { PDFName, PDFRef, PrivateConstructorError } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PrivateConstructorError } from 'src/core/errors.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 import { security } from './shared.js';
 
 describe(`PDFName`, () => {
