+++ tests/core/objects/PDFNull.spec.ts
@@ -1,5 +1,7 @@
-import { PDFNull, PDFRef } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PDFNull } from 'src/core/objects/PDFNull.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 import { security } from './shared.js';
 
 describe(`PDFNull`, () => {
