+++ tests/core/objects/PDFNumber.spec.ts
@@ -1,5 +1,7 @@
-import { PDFNumber, PDFRef } from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PDFNumber } from 'src/core/objects/PDFNumber.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 import { security } from './shared.js';
 
 describe(`PDFNumber`, () => {
