+++ tests/core/operators/PDFOperator.spec.ts
@@ -1,10 +1,9 @@
-import {
-  PDFNumber,
-  PDFOperator,
-  PDFOperatorNames as Ops,
-  PDFString,
-} from 'src/core/index.js';
-import { toCharCode, typedArrayFor } from 'src/utils/index.js';
+import { PDFNumber } from 'src/core/objects/PDFNumber.js';
+import { PDFString } from 'src/core/objects/PDFString.js';
+import { PDFOperator } from 'src/core/operators/PDFOperator.js';
+import { PDFOperatorNames as Ops } from 'src/core/operators/PDFOperatorNames.js';
+import { typedArrayFor } from 'src/utils/arrays.js';
+import { toCharCode } from 'src/utils/strings.js';
 
 describe(`PDFOperator`, () => {
   it(`can be constructed with args`, () => {
