+++ tests/core/parser/PDFObjectStreamParser.spec.ts
@@ -1,18 +1,16 @@
-import {
-  PDFArray,
-  PDFBool,
-  PDFContext,
-  PDFDict,
-  PDFHexString,
-  PDFName,
-  PDFNull,
-  PDFNumber,
-  PDFObjectStreamParser,
-  PDFRawStream,
-  PDFRef,
-  PDFString,
-  ReparseError,
-} from 'src/core/index.js';
+import { ReparseError } from 'src/core/errors.js';
+import { PDFContext } from 'src/core/PDFContext.js';
+import { PDFArray } from 'src/core/objects/PDFArray.js';
+import { PDFBool } from 'src/core/objects/PDFBool.js';
+import { PDFDict } from 'src/core/objects/PDFDict.js';
+import { PDFHexString } from 'src/core/objects/PDFHexString.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFNull } from 'src/core/objects/PDFNull.js';
+import { PDFNumber } from 'src/core/objects/PDFNumber.js';
+import { PDFRawStream } from 'src/core/objects/PDFRawStream.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { PDFString } from 'src/core/objects/PDFString.js';
+import { PDFObjectStreamParser } from 'src/core/parser/PDFObjectStreamParser.js';
 import { readBinaryFileSync } from '../../test-utils.js';
 
 const readData = (file: string) =>
