+++ tests/core/structures/PDFPageLeaf.spec.ts
@@ -1,12 +1,10 @@
-import {
-  PDFArray,
-  PDFContext,
-  PDFDict,
-  PDFName,
-  PDFPageLeaf,
-  PDFPageTree,
-  PDFRef,
-} from 'src/core/index.js';
+import { PDFContext } from 'src/core/PDFContext.js';
+import { PDFArray } from 'src/core/objects/PDFArray.js';
+import { PDFDict } from 'src/core/objects/PDFDict.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { PDFPageLeaf } from 'src/core/structures/PDFPageLeaf.js';
+import { PDFPageTree } from 'src/core/structures/PDFPageTree.js';
 
 describe(`PDFPageLeaf`, () => {
   it(`can be constructed directly from a Map and PDFContext`, () => {
