+++ tests/utils/pdfDocEncoding.spec.ts
@@ -1,4 +1,5 @@
-import { range, pdfDocEncodingDecode } from 'src/utils/index.js';
+import { range } from 'src/utils/arrays.js';
+import { pdfDocEncodingDecode } from 'src/utils/pdfDocEncoding.js';
 
 type Mapping = [number, string];
